requirements.txt
pandas
pysentimiento
matplotlib
seaborn
wordcloud
python-dotenv
google-api-python-client